refactor(lens): consolidate formatters into lib/glens/formatters.ts
fmtNumber, fmtCost, fmtDate, relativeTime previously duplicated across
GLensPageBubble, GenericTableBubble, and GlensDashboard (three slightly
different fmtNumber impls). Consolidated into a single lib. Canonical
fmtNumber uses Math.abs + strips trailing .0 (superset of the three
prior variants).

Behavior-preserving except GlensPageBubble.fmtNumber now handles
negatives + strips .0 — visual improvement, no callers rely on the old
output.

TypeScript check passes.

Refactor #2 of 3 from the Lens modularity audit. -47 LOC across three
components; +36 in the new shared lib.
